refactoring exercise answer
